commit 1: первые настройки, накидал основные компоненты
