+++ package.json
@@ -9,7 +9,15 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-scripts": "4.0.3",
-    "web-vitals": "^1.0.1"
+    "web-vitals": "^1.0.1",
+    "@types/node": "^16.11.12",
+    "@types/react": "^17.0.37",
+    "@types/react-dom": "^17.0.11",
+    "@types/react-router-dom": "^5.3.2",
+    "classnames": "^2.3.1",
+    "gh-pages": "^3.2.3",
+    "react-router-dom": "^6.0.2",
+    "typescript": "^4.5.2"
   },
   "scripts": {
     "start": "react-scripts start",
