+++ src/components/Footer/Footer.tsx
@@ -1,0 +1,5 @@
+import './Footer.css';
+
+export const Footer = () => (
+  <div>I-Novus GRPI 2021</div>
+);
