+++ src/components/Header/Header.tsx
@@ -1,0 +1,5 @@
+import './Header.css';
+
+export const Header = () => (
+  <div>QUIZZ!</div>
+);
