+++ src/components/Main/Main.css
@@ -1,0 +1,8 @@
+.content {
+  width: 100%;
+  padding: 1rem 0;
+  min-height: 80vh;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
