+++ src/components/Main/Main.tsx
@@ -1,0 +1,8 @@
+import './Main.css';
+
+export const Main = () => (
+  <main className="content">
+    main
+  </main>
+);
+
